refactor: preparations for add role flow
